+++ packages/maps/src/map-utils/google-map-utils.ts
@@ -314,6 +314,16 @@ function setDivMarkerStyle(item, options) {
   if (anchor) item.setAnchor(anchor);
 }
 
+function getHeatmap(options = {}) {
+  let heatmap = new google.maps.visualization.HeatmapLayer({data: [], ...options});
+  return heatmap;
+}
+
+function updateHeatMapLatlng(latlngs, heatmap) {
+  let latlngsPos = latlngs.map((pos: HtPosition) => GetLatlng(pos));
+  heatmap.setData(latlngsPos)
+}
+
 export const GoogleMapUtils: MapUtils = {
   type: "google",
   setMap: SetMap,
@@ -350,5 +360,7 @@ export const GoogleMapUtils: MapUtils = {
   setDivContent,
   getDivMarker,
   setDivMarkerStyle,
-  setPathPositionTimeArray
+  setPathPositionTimeArray,
+  getHeatmap,
+  updateHeatMapLatlng
 };